GraphTotalTransform only in matrix, Test Scene with add Planet and Satellites

+++ CustomEngine/src/Engine/Camera.cpp
@@ -52,7 +52,7 @@ Camera::~Camera()
 
 void Camera::Update(float deltatime) {
 
-	SimulateWalking(0.015f, glfwGetTime());
+	SimulateWalking(0.015f, static_cast<float>(glfwGetTime()));
 	movements.move(this, deltatime);
 
 };
